upload_revision.py: Check before trying to rmtree RUST_SRC_PATH
Bug: 1401042
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4166074
Commit-Queue: Arthur Eubanks <aeubanks@google.com>
Auto-Submit: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1093110}

+++ tools/clang/scripts/upload_revision.py
@@ -206,8 +206,9 @@ def GetLatestRustVersion(sub_revision: int):
 
 
 def FetchRust(rust_version):
-  print(f'Removing {RUST_SRC_PATH}.')
-  shutil.rmtree(RUST_SRC_PATH)
+  if os.path.exists(RUST_SRC_PATH):
+    print(f'Removing {RUST_SRC_PATH}.')
+    shutil.rmtree(RUST_SRC_PATH)
 
   print(f'Fetching Rust {rust_version} into {RUST_SRC_PATH}')
   version_str = rust_version.string_with_dashes(with_tag=True)
